Memory usage decrease and fixed buffer overflow

+++ Dijsktra/Graph.hpp
@@ -16,15 +16,15 @@ class Graph
         int getShortestPathLen();
         
         Node adjList[64][8]{};
-        int adjListSize[64]{}; // Keeps track of each adjacency list size (the nested array)
-        int shortestPath[64];   // Path of nodes to traverse to get from source node to destination node.
-        int distArray[64]{}; // Track distance values from starting node
-        int prevArray[64]{}; // Tracks the index of the node you took to get to some node i
+        short adjListSize[64]{}; // Keeps track of each adjacency list size (the nested array)
+        short shortestPath[64];   // Path of nodes to traverse to get from source node to destination node.
+        unsigned short distArray[64]{}; // Track distance values from starting node
+        short prevArray[64]{}; // Tracks the index of the node you took to get to some node i
 
     private:
         // Index of array will correspond to origin node. Nested array will contain adjacent nodes
-        int shortestPathLen = 0;
-        int numOfNodes = 0;
+        short shortestPathLen = 0;
+        short numOfNodes = 0;
         //std::vector<minHeapNode> MSToutput = std::vector<minHeapNode>();
 };
